suppress warnings
git-svn-id: https://svn.salilab.org/imp/trunk@8483 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/atom/src/pdb.cpp
@@ -462,7 +462,7 @@ void write_pdb(const Particles& ps, TextOutput out)
         << std::endl;*/
       }
       out.get_stream() << get_pdb_string(core::XYZ(ps[i]).get_coordinates(),
-                                     use_input_index? ad.get_input_index(): i+1,
+               use_input_index? ad.get_input_index(): static_cast<int>(i+1),
                                      ad.get_atom_type(),
                                      rd.get_residue_type(),
                                      chain,
